Added note for any future users looking at demo
Also summarizing here: this was an idea I had and I wanted to clarify that I don't necessarily expect it to happen, as I myself am having trouble thinking of a way to do it as well. It may even be best as an animation walkthrough sort of thing, rather than literally an online application on the website.