Add files via upload
We made the Online Demo Page. Basing it off the video, we created a sample as to what RaiderPlanner will look like for users. This includes a class list with a progress image, and for each class a chart documenting the assignments they have for the class. This extension will give users a feel of how RaiderPlanner will work for them in a almost tutorial like manner.

Sam Daniel 
Jordan Bintz